+++ src/lib/pclsync/pssl.h
@@ -75,7 +75,7 @@ int psync_ssl_connect(psync_socket_t sock, void **sslconn, const char *hostname)
 int psync_ssl_connect_finish(void *sslconn, const char *hostname);
 void psync_ssl_free(void *sslconn);
 int psync_ssl_shutdown(void *sslconn);
-int psync_ssl_pendingdata(void *sslconn);
+size_t psync_ssl_pendingdata(void *sslconn);
 int psync_ssl_read(void *sslconn, void *buf, int num);
 int psync_ssl_write(void *sslconn, const void *buf, int num);
